rm simulation, move ctf_utils
